requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
email_validator==2.1.0  # Required for WTForms email validation
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
psycopg2-binary==2.9.9  # PostgreSQL adapter for Python
pyodbc==4.0.39  # Microsoft SQL Server driver
sqlalchemy-pytds==0.3.2  # Alternative to sqlalchemy-pyodbc-azure for SQL Server

# Azure Key Vault integration for secure credential management
azure-identity==1.14.0  # DefaultAzureCredential for authentication
azure-keyvault-secrets==4.7.0  # Azure Key Vault Secrets client

# Removed all Azure CLI and identity packages to resolve deployment conflicts.
# These packages aren't needed for the Flask application to run on Azure App Service.
# GitHub Actions handles the deployment process without requiring these dependencies.